modules/odr_caption/requirements.txt
vllm
transformers
qwen-vl-utils[decord]
fastapi
click
openai
pydantic
termcolor
pillow
python-dotenv
outlines
torch
torchvision
httpx
accelerate
termcolor
python-multipart
